Refactor: Change directory proma backend init
