+++ proma/src/main/java/ai/platform/proma/config/WebConfig.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.configuration;
+package ai.platform.proma.config;
 
 
 import lombok.RequiredArgsConstructor;
