+++ proma/src/main/java/ai/platform/proma/dto/exception/JSONConvertExceptionDto.java
