+++ proma/src/main/java/ai/platform/proma/repositroy/UserRepository.java
@@ -3,8 +3,6 @@ package ai.platform.proma.repositroy;
 import ai.platform.proma.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface UserRepository extends JpaRepository<User, Long> {
 
 }
